refactor: ♻️ saveXXX to updateUserSelfCareTopics
