+++ libs/modules/users/application/src/index.ts
@@ -2,4 +2,4 @@ export * from './lib/users.service';
 export * from './lib/users-service-factory';
 
 export * from './lib/get-user-by-id.use-case';
-export * from './lib/save-user-self-care-topics.use-case';
+export * from './lib/update-user-self-care-topics.use-case';
