+++ libs/modules/users/application/src/lib/update-user-self-care-topics.use-case.ts
@@ -2,9 +2,9 @@ import { SelfCareTopic } from '@self-care-topics/domain';
 
 import { UsersService } from './users.service';
 
-export const saveUserSelfCareTopicsUseCase = async (
+export const updateUserSelfCareTopicsUseCase = async (
   topics: SelfCareTopic[],
   usersService: UsersService,
 ): Promise<void> => {
-  return await usersService.saveUserSelfCareTopics(topics);
+  return await usersService.updateUserSelfCareTopics(topics);
 };
